Fixes codes for "Bodo".
As reported in #15, we have confused two languages called "Bodo". The
data previously labeled `boy` is actually `brx`, a Sino-Tibeto-Burman
language of India, not `boy`, a Niger-Congo language spoken in Africa.

(Both can be called "Bodo"; `brx` is also know as "Boro").

I have moved the .tsv file and fixed the ISO 639 and Glottolog codes `codes.tsv` to reflect.

Closes #15.